chore(amp): 更新 Rockchip HAL SDK
同步新版 Rockchip HAL SDK，补齐 middleware/rpmsg-lite、OpenAMP、sdhci 等中间件源码，并新增 RK2118、RK3506、RK3576、RV1103B 等 SoC 支持。

移除旧的 RK3568 rpmsg-lite-port 临时平台文件，改用 HAL SDK 自带的 middleware/rpmsg-lite 平台实现。

+++ components/amp/rockchip/hal/lib/CMSIS/Device/RV1106/Source/Templates/system_rv1106_mcu.c
@@ -20,7 +20,7 @@ static void BssInit(void)
     size = (__BSS_END__ - __BSS_START__ + 1) * 4;
 
     memset(p, 0x0, size);
-    HAL_DCACHE_CleanByRange((uint32_t)p, size);
+    HAL_DCACHE_CleanByRange((uintptr_t)p, size);
 }
 
 /*----------------------------------------------------------------------------
